moved light dist approx function to cob_light driver
